requirements.txt
Flask
python-dotenv
google-genai
Pillow
werkzeug
google-cloud-storage